+++ crates/data-connectors/connector-ducklake/Cargo.toml
@@ -1,0 +1,37 @@
+##
+## Copyright 2026 The Spice.ai OSS Authors
+##
+## Licensed under the Apache License, Version 2.0 (the "License");
+## you may not use this file except in compliance with the License.
+## You may obtain a copy of the License at
+##
+##      https://www.apache.org/licenses/LICENSE-2.0
+##
+## Unless required by applicable law or agreed to in writing, software
+## distributed under the License is distributed on an "AS IS" BASIS,
+## WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+## See the License for the specific language governing permissions and
+## limitations under the License.
+##
+
+[package]
+name = "connector-ducklake"
+edition.workspace = true
+version.workspace = true
+rust-version.workspace = true
+license-file.workspace = true
+description = "DuckLake data connector for Spice.ai runtime"
+
+[dependencies]
+async-trait.workspace = true
+data_components = { path = "../../data_components", features = ["duckdb"] }
+datafusion.workspace = true
+datafusion-table-providers = { workspace = true, features = ["duckdb"] }
+duckdb = { workspace = true, features = ["bundled", "r2d2", "vtab", "vtab-arrow", "appender-arrow"] }
+runtime = { path = "../../runtime" }
+runtime-datafusion = { path = "../../runtime-datafusion" }
+snafu.workspace = true
+tokio.workspace = true
+
+[features]
+default = []
