+++ crates/data-connectors/connector-github/Cargo.toml
@@ -1,0 +1,55 @@
+##
+## Copyright 2026 The Spice.ai OSS Authors
+##
+## Licensed under the Apache License, Version 2.0 (the "License");
+## you may not use this file except in compliance with the License.
+## You may obtain a copy of the License at
+##
+##      https://www.apache.org/licenses/LICENSE-2.0
+##
+## Unless required by applicable law or agreed to in writing, software
+## distributed under the License is distributed on an "AS IS" BASIS,
+## WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+## See the License for the specific language governing permissions and
+## limitations under the License.
+##
+
+[package]
+name = "connector-github"
+edition.workspace = true
+version.workspace = true
+rust-version.workspace = true
+license-file.workspace = true
+description = "GitHub data connector for Spice.ai runtime"
+
+[dependencies]
+app = { path = "../../../crates/app" }
+arrow.workspace = true
+arrow-schema.workspace = true
+arrow_tools = { path = "../../arrow_tools" }
+async-trait.workspace = true
+chrono.workspace = true
+data_components = { path = "../../data_components" }
+datafusion.workspace = true
+futures.workspace = true
+globset.workspace = true
+governor.workspace = true
+graphql-parser.workspace = true
+http.workspace = true
+opentelemetry.workspace = true
+reqwest.workspace = true
+runtime = { path = "../../runtime" }
+runtime-datafusion = { path = "../../runtime-datafusion" }
+runtime-parameters = { path = "../../runtime-parameters" }
+runtime-rate-control = { path = "../../runtime-rate-control" }
+secrecy.workspace = true
+serde_json.workspace = true
+snafu.workspace = true
+token_provider = { path = "../../token_provider" }
+tokio.workspace = true
+tracing.workspace = true
+url.workspace = true
+util = { path = "../../util" }
+
+[features]
+default = []
