+++ crates/data-connectors/connector-github/src/projects.rs
@@ -14,11 +14,11 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-use crate::dataconnector::ConnectorComponent;
+use runtime::dataconnector::ConnectorComponent;
 
-use super::{GitHubTableArgs, GitHubTableGraphQLParams};
+use crate::{GitHubTableArgs, GitHubTableGraphQLParams};
 use arrow_schema::{DataType, Field, Schema, SchemaRef};
-use data_components::graphql::{ErrorChecker, GraphQLContext, client::UnnestBehavior};
+use data_components::graphql::{Error, ErrorChecker, GraphQLContext, client::UnnestBehavior};
 use http::{HeaderMap, HeaderValue};
 use serde_json::Value;
 use std::sync::Arc;
@@ -75,9 +75,11 @@ impl GraphQLContext for ProjectsTableArgs {
                                 tracing::error!(
                                     "GitHub returned a misleading projects error for {target}; treating it as a permissions failure"
                                 );
-                                return Err(data_components::graphql::Error::InvalidCredentialsOrPermissions {
-                                message: format!("Failed to access {target_kind} for {target}: GitHub reported an internal query error, which usually means the GitHub App lacks project read permissions. Verify the app has the required project access."),
-                            });
+                                return Err(Error::InvalidCredentialsOrPermissions {
+                                    message: format!(
+                                        "Failed to access {target_kind} for {target}: GitHub reported an internal query error, which usually means the GitHub App lacks project read permissions. Verify the app has the required project access."
+                                    ),
+                                });
                             }
                         }
                     }
