+++ crates/data-connectors/connector-github/src/workflows.rs
@@ -14,7 +14,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-use crate::component::dataset::Dataset;
 use async_trait::async_trait;
 use data_components::github::{Error as GithubError, GithubRestClient};
 use datafusion::{
@@ -41,11 +40,12 @@ use datafusion::{
     },
 };
 use futures::{TryFutureExt, TryStreamExt};
+use runtime::component::dataset::Dataset;
 use std::{any::Any, sync::Arc};
 
 use arrow::datatypes::{DataType, Field, Schema, SchemaRef, TimeUnit};
 
-use super::ConnectorComponent;
+use runtime::dataconnector::ConnectorComponent;
 
 #[derive(Debug)]
 pub struct WorkflowsTableProvider {
@@ -61,7 +61,7 @@ impl WorkflowsTableProvider {
         owner: &str,
         repo: &str,
         dataset: &Dataset,
-    ) -> crate::dataconnector::DataConnectorResult<Self> {
+    ) -> runtime::dataconnector::DataConnectorResult<Self> {
         let fields = vec![
             Field::new("id", DataType::Int64, false),
             Field::new("name", DataType::Utf8, false),
@@ -92,7 +92,7 @@ impl WorkflowsTableProvider {
             if e.downcast_ref::<GithubError>()
                 .is_some_and(|err| matches!(err, GithubError::RateLimited { .. }))
             {
-                return Err(super::DataConnectorError::RateLimited {
+                return Err(runtime::dataconnector::DataConnectorError::RateLimited {
                     dataconnector: "github".to_string(),
                     connector_component: component,
                     source: e,
@@ -105,11 +105,13 @@ impl WorkflowsTableProvider {
                     "GitHub workflows provider initialization for {component} could not validate access because GitHub is temporarily unavailable: {e} The dataset will retry on the next query or refresh."
                 );
             } else {
-                return Err(super::DataConnectorError::UnableToGetReadProvider {
-                    dataconnector: "github".to_string(),
-                    connector_component: component,
-                    source: e,
-                });
+                return Err(
+                    runtime::dataconnector::DataConnectorError::UnableToGetReadProvider {
+                        dataconnector: "github".to_string(),
+                        connector_component: component,
+                        source: e,
+                    },
+                );
             }
         }
