+++ crates/data-connectors/connector-spiceai/Cargo.toml
@@ -1,0 +1,46 @@
+##
+## Copyright 2026 The Spice.ai OSS Authors
+##
+## Licensed under the Apache License, Version 2.0 (the "License");
+## you may not use this file except in compliance with the License.
+## You may obtain a copy of the License at
+##
+##      https://www.apache.org/licenses/LICENSE-2.0
+##
+## Unless required by applicable law or agreed to in writing, software
+## distributed under the License is distributed on an "AS IS" BASIS,
+## WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+## See the License for the specific language governing permissions and
+## limitations under the License.
+##
+
+[package]
+name = "connector-spiceai"
+edition.workspace = true
+version.workspace = true
+rust-version.workspace = true
+license-file.workspace = true
+description = "Spice.ai platform data connector for Spice.ai runtime"
+
+[dependencies]
+arrow-flight.workspace = true
+linkme.workspace = true
+async-stream.workspace = true
+async-trait.workspace = true
+data_components = { path = "../../data_components" }
+datafusion.workspace = true
+datafusion-federation.workspace = true
+flight_client = { path = "../../flight_client" }
+futures.workspace = true
+ns_lookup = { path = "../../ns_lookup" }
+runtime = { path = "../../runtime" }
+secrecy.workspace = true
+snafu.workspace = true
+spice-cloud-client.workspace = true
+spiceai-connector-types = { path = "../../spiceai-connector-types" }
+tokio.workspace = true
+tonic.workspace = true
+tracing.workspace = true
+
+[features]
+default = []
