+++ crates/spiceai-connector-types/Cargo.toml
@@ -1,0 +1,33 @@
+##
+## Copyright 2026 The Spice.ai OSS Authors
+##
+## Licensed under the Apache License, Version 2.0 (the "License");
+## you may not use this file except in compliance with the License.
+## You may obtain a copy of the License at
+##
+##      https://www.apache.org/licenses/LICENSE-2.0
+##
+## Unless required by applicable law or agreed to in writing, software
+## distributed under the License is distributed on an "AS IS" BASIS,
+## WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+## See the License for the specific language governing permissions and
+## limitations under the License.
+##
+
+[package]
+name = "spiceai-connector-types"
+edition.workspace = true
+version.workspace = true
+rust-version.workspace = true
+license-file.workspace = true
+description = "Shared types for the Spice.ai data connector, used by both the connector crate and the runtime catalog connector"
+
+[dependencies]
+data_components = { path = "../data_components" }
+datafusion.workspace = true
+flight_client = { path = "../flight_client" }
+snafu.workspace = true
+tonic.workspace = true
+
+[features]
+default = []
